// src/app/shared/store/movies/types.ts
export enum MoviesActionTypes {
    LoadFilms = '[Films] Load All Films',
    LoadFilmsSuccess = '[Films] Load Films Success',
    LoadFilmsFailure = '[Films] Load Films Failure',
    LoadPeople = '[People] Load People',
    LoadPeopleSuccess = '[People] Load People Success',
    LoadPeopleFailure = '[People] Load People Failure',
    LoadSpecies = '[Species] Load Species',
    LoadSpeciesSuccess = '[Species] Load Species Success',
    LoadSpeciesFailure = '[Species] Load Species Failure',
}